match Host headers containing port
The change to protoHandler.ServeHTTP is a hack to work around
https://golang.org/issue/10463

Some clients (like curl when given a user-provided Host header with
-H'Host: foobar:443) will always include the 443 port in the Host header
of their HTTPS requests. They will do so even in the face of a
redirect's Location that does not include a port.

On the other side, net/http.ServeMux does not match against ports
when they are included as part of mux.Handle("host:port", doSomething).

This also means that we never had to be smart at boot time about
dropping the port if it wasn't 443.

Closes #25