add ability to only allow certain origins in API
If a file path to an allowed origins config file is given, requests with
a host in the Origin or Referer header that is not in the Allowed list are
given a 429 status code error. Requests without a Referer or Origin
header will continue to work as the API is meant to be used to test
client tools.

This makes the API work as documented. The API is not meant for testing
other sites' users' browsers for free. Pricing information for sites
that would like to use the API to test their users' browsers is
available by contacting jeff@darkishgreen.com