add optional Google Cloud Logging of allowances (#104)
This logs data about why an API request was allowed or disallowed to Google Cloud Logging. It's optional and, if not configured, howsmyssl will run just fine.

Brings in a bunch of dependencies to pull this off. Had to fix up travis to prevent the vendored code from breaking the build.