update Google Cloud import paths (#128)
This moves from the old Google Cloud package import paths to the new ones as
laid out in˘ #124

This is unfortunately extremely large. The new import paths required code only
available in newer checkouts of google.golang.org/api. In order to avoid build
problems, I went ahead and refreshed all of the dependencies that were only
vendored because of the Google Cloud packages.

This refresh, unfortunately, include golang.org/x/net/http2/h2demo which
includes a bunch of camilstore.org packages that are otherwise not used in
howsmyssl.

We could complain about the packaged not being behind a testdata dir, or we
could try to argue that vendoring tools should detect unused deps and not
include them (I think there's some strong problems against this latter
solution), but we still have to live with it for now.

Fixes #124