support TLS 1.3 supported_versions extension
Recent versions of the TLS 1.3 draft use a TLS extension called
supported_versions to announce the versions supported by the client. And
TLS 1.3 clients that support different drafts of the TLS 1.3 specs send
down different versions with the spec draft number bitmasked in.

This patch adds support for all of that behavior.

Fixes #201.