use allowlists instead of the origins file (#208)
This new config file is fed by the subscription service and allows us to more
easily handle cancellation and other issues.

We check the exact domain allowlist, if the given domain in there, we return
good.

We check the allowlist that has domains that have their subdomains (but
not necessarily that domain) allowed, and if the given domain is in there, we
return good.

We check the blocklists, and if the given domain is in there or if the given
domain is a subdomain of a domain that's in there, we we return bad.  Else, we
return good.

This lets us handle domains that flip flop between subscribed and not subscribed
easily while also allowing for holes to be punched in wildcard blocks by
subscriptions and for blocks to happen ahead of a subscription.