return 200 for blocked JSONP requests (#214)
Turns out, when you do a <script src="yadda?callback=parseTLSThing">, the script
won't run if a 4xx error code is returned. So, in order for folks to show errors
to their users, and not just quietly log in their console, we have to return 200
OK even on blocked domains.

To be clear, normal AJAX requests would work, but people do that script tag
thing to be JSONP-y.

This patch also includes some new tests (and, along the way, a refactoring of
the current ones).