update Go outside of dockerfile (#364)
Update the Go version used in the GitHub workflows. We may want to reconsider moving this back to the workflows now that some auto-update systems can update workflows

ECDSA and RSA ciphersuites have switch places in Go's ClientHello in the intervening versions, so the tests needed a fix.

Also, bump the `go.mod` go versions and do the re-vendoring it requires.