replace use of Hijack with ConnContext
This patch removes the use of `http.Hijacker` to pass around the extra
TLS client information and replaces it with `http.Server.ConnContext`.

In order to get all of the information howsmyssl needed on the TLS
client connected to it, it needed to get access to the TLS connection
for the incoming HTTP request. And howmyssl needed a fork of
`crypto/tls` to do gather all the information it needs from the TLS
client.

This, if I remember correctly, predated the existence of the
`http.Request.TLS` field, but even that field was insufficient. It's
only filled when the the `net.Listener` the `http.Server` is on is
returns exactly a `crypto/tls.Conn`, and our fork of `crypto/tls` means
that isn't ever true.

But, in the meantime, a new feature was added to `http.Server`: the
`ConnContext`. By using it to set our own `conn` type on the context, we
make our connection available to our HTTP handlers without the need for
`Hijack`.

Not needing `Hijack` means that we have to do significantly less
connection management, getting to use more standard `net/http` handler
tooling, and avoid having to do all those HTTP response formatting work
by hand.

We still have to do some amount of connection managagement in the form
of setting `Connection: close` to prevent HTTP keepalives. But all told,
the amount of code and mental overhead is significantly reduced.

We still don't add HTTP/2 support and likely won't til we solve at least
issue jmhodges/howmyssl#526 and also would still have to ratchet the max
streams down to 1 per connection to make sure that's safe.

This also allow us to add compression to our APIs per #192

Fixes #528