remove port stripping hack from our handlers
In response to #25, we added this hack to strip the port of the Host
header before passing it to our virtual host redirecting ServeMux. I
also filed https://github.com/golang/go/issues/10463 and that was fixed
in 2017 and released soon after.

So, since we're on a modern version of Go, this code is redundant and
can be removed.