pass loggers in and use structured logging
This is to both quiet the tests by passing in a logger that understands
`testing` to the code that can be chatty and to enable more
understandable and usable logs structured logging.

Go 1.21 comes out next month and our new dep on exp/slog will be able to
become a dependency on the stdlib version of slog instead.