remove use of grpc.WithBlock in Cloud Logging (#720)
The GRPC does say not to use this. They also go on to say we're making
the wrong assumption that our configuration will be wrong more than
we'll have connectivity issues we could ignore.

The former is fine, but I suspect the latter is liekly not for us.

The docs describe how to use `ClientConn.GetState` and `Connect` to
reproduce this behavior, but the `logging` client doesn't immediately
give us access to the `ClientConn`. It does, however, have a higher
level `Ping` method that is not just a connectivity check but a
permissions check, as well. Totally fine and perhaps even better than
what was there.